Eddy covariance data from Gebesee crop site, Germany

The data frame 'DEGebExample' contains half-hourly eddy covariance measurements from Gebesee of the years 2004 to 2006.

data(DEGebExample)

Format

For each column, the attributes 'varnames' for the variable names and 'units' for the variable units are provided.

Time stamp

DateTime: POSIXct-time of the end of the half-hour period, Use as.POSIXlt(DateTime)$year to get hour, day of year, ...

Flux measurements

NEE

Meteo measurements

Rg, Tair, rH, VPD, Ustar

For processing of the example data see vignette("DEGebExample").

Details

DISCLAIMER: This example dataset should only be used for test purposes of the REddyProc R package. For other uses, the data is openly available through the European Fluxes Database (http://www.europe-fluxdata.eu/home/site-details?id=3) and upon registration the current version can be downloaded there.

Source

The data was downloaded from http://www.europe-fluxdata.eu at date 2016-01-25.