Gradient of the Light Response Function

Gradient of the Light Response Function

LRC_computeLRCGradient(
  theta,
  Rg,
  VPD,
  Temp,
  VPD0 = 10,
  fixVPD = (k == 0),
  TRef = 15
)

Arguments

theta

numeric vector of parameters. If theta is a matrix, a different row of parameters is used for different entries of other inputs.

Rg

ppfd -> photosynthetic flux density mumol / m2 / s or Global Radiation

VPD

VPD -> Vapor Pressure Deficit hPa

Temp

Temp -> Temperature degC

VPD0

VPD0 hPa -> Parameters VPD0 fixed to 10 hPa according to Lasslop et al 2010

fixVPD

if TRUE the VPD effect is not considered and VPD is not part of the computation

TRef

numeric scalar of Temperature (degree Celsius) for reference respiration RRef

Value

list with gradient matrices. For each record (length(Rg)), c("k", "beta", "alpha", "RRef")