Optimize parameters with refitting with some fixed parameters if outside bounds

Optimize parameters with refitting with some fixed parameters if outside bounds

LRC_optimLRCBounds(
  theta0,
  parameterPrior,
  ...,
  dsDay,
  lastGoodParameters,
  ctrl
)

Arguments

theta0

numeric vector of initial parameter estimate

parameterPrior

numeric vector of prior estimate of model parameters

...

further parameters to .optimLRC

dsDay

dataframe of NEE, sdNEE and predictors Rg, VPD and Temp

lastGoodParameters

numeric vector of last successful fit

ctrl

list of further controls, such as isNeglectVPDEffect = TRUE

Value

list of result of LRC_optimLRCOnAdjustedPrior()

theta: vector of optimized parameters

iOpt: vector of positions of parameters that are optimized

thetaInitialGuess: initial guess from data see LRC_fitLRC()

Details

If parameters alpha or k are outside bounds (Table A1 in Lasslop 2010), refit with some parameters fixed to values from fit of previous window.

See also