Class "LightResponseCurveFitter"

Base class for fitting parameters to light response curves (LRC)

Concrete classes for the following LRC functions are available:

common rectangular hyperbolic light-response: RectangularLRCFitter-class

nonrectangular hyperbolic light-response: NonrectangularLRCFitter-class

logistic sigmoid light-response: LogisticSigmoidLRCFitter-class

They mostly differ in their prediction of GPP by method LightResponseCurveFitter_predictGPP.

Extends

All reference classes extend and inherit methods from "envRefClass".

Author

TW

Methods

LightResponseCurveFitter_computeLRCGradient(theta, Rg, VPD, Temp, VPD0, fixVPD, TRef):

LightResponseCurveFitter_predictGPP(Rg, ...):

LightResponseCurveFitter_predictLRC(theta, Rg, VPD, Temp, VPD0, fixVPD, TRef):

LightResponseCurveFitter_computeCost(thetaOpt, theta, iOpt, flux, sdFlux, parameterPrior, sdParameterPrior, ...):

LightResponseCurveFitter_optimLRC(theta, iOpt, sdParameterPrior, ..., ctrl, isUsingHessian):

LightResponseCurveFitter_isParameterInBounds(theta, sdTheta, RRefNight, ctrl):

LightResponseCurveFitter_optimLRCOnAdjustedPrior(theta, iOpt, dsDay, parameterPrior, ctrl, ...):

LightResponseCurveFitter_getOptimizedParameterPositions(isUsingFixedVPD, isUsingFixedAlpha):

LightResponseCurveFitter_optimLRCBounds(theta0, parameterPrior, ..., lastGoodParameters, ctrl):

LightResponseCurveFitter_getParameterInitials(thetaPrior):

LightResponseCurveFitter_getPriorScale(thetaPrior, medianRelFluxUncertainty, nRec, ctrl):

LightResponseCurveFitter_getPriorLocation(NEEDay, RRefNight, E0):

LightResponseCurveFitter_fitLRC(dsDay, E0, sdE0, RRefNight, controlGLPart, lastGoodParameters):

LightResponseCurveFitter_getParameterNames():