Post Processing of (Half-)Hourly Eddy-Covariance Measurements

Standard and extensible Eddy-Covariance data post-processing including uStar-filtering, gap-filling, and flux-partitioning (Wutzler et al. (2018) <doi:10.5194/bg-15-5015-2018>).

The Eddy-Covariance (EC) micrometeorological technique quantifies continuous exchange fluxes of gases, energy, and momentum between an ecosystem and the atmosphere. It is important for understanding ecosystem dynamics and upscaling exchange fluxes. (Aubinet et al. (2012) <doi:10.1007/978-94-007-2351-1>).

This package inputs pre-processed (half-)hourly data and supports further processing. First, a quality-check and filtering is performed based on the relationship between measured flux and friction velocity (uStar) to discard biased data (Papale et al. (2006) <doi:10.5194/bg-3-571-2006>).

Second, gaps in the data are filled based on information from environmental conditions (Reichstein et al. (2005) <doi:10.1111/j.1365-2486.2005.001002.x>).

Third, the net flux of carbon dioxide is partitioned into its gross fluxes in and out of the ecosystem by night-time based and day-time based approaches (Lasslop et al. (2010) <doi:10.1111/j.1365-2486.2009.02041.x>).

A general description and an online tool based on this package can be found here: https://www.bgc-jena.mpg.de/bgi/index.php/Services/REddyProcWeb.

Details

A detailed example of the processing can be found in the useCase vignette.

A first overview of the REddyProc functions:

These functions help with the preparation of your data for the analysis:

Loading text files into dataframes: fLoadTXTIntoDataframe

Preparing a proper time stamp: help_DateTimes

Calculating latent variables, e.g. VPD: fCalcVPDfromRHandTair

Then the data can be processed with the sEddyProc-class R5 reference class:

Initializing the R5 reference class: sEddyProc_initialize

Estimating the turbulence criterion, Ustar threshold, for omitting data from periods of low turbulence: Functions sEddyProc_sEstUstarThreshold and sEddyProc_sEstUstarThresholdDistribution .

Gap filling: sEddyProc_sMDSGapFill and sEddyProc_sMDSGapFillAfterUstar.

Flux partitioning based on Night-Time: sEddyProc_sMRFluxPartition

Flux partitioning based on Day-Time: sEddyProc_sGLFluxPartition

Processing across different scenarios of u* threshold estimate is supported by

Estimating the turbulence criterion, Ustar threshold, for omitting data from periods of low turbulence: sEddyProc_sEstimateUstarScenarios and associated

query the thresholds to be used sEddyProc_sGetUstarScenarios

set the thresholds to be used sEddyProc_sSetUstarScenarios

query the estimated thresholds all different aggregation levels sEddyProc_sGetEstimatedUstarThresholdDistribution

Gap-Filling: sEddyProc_sMDSGapFillUStarScens

Flux partitioning based on Night-Time (Reichstein 2005): sEddyProc_sMRFluxPartitionUStarScens

Flux partitioning based on Day-Time (Lasslop 2010): sEddyProc_sGLFluxPartitionUStarScens

Flux partitioning based on modified Day-Time (Keenan 2019): sEddyProc_sTKFluxPartitionUStarScens

Before or after processing, the data can be plotted:

Fingerprint: sEddyProc_sPlotFingerprint

Half-hourly fluxes and their daily means: sEddyProc_sPlotHHFluxes

Daily sums (and their uncertainties): sEddyProc_sPlotDailySums

Diurnal cycle: sEddyProc_sPlotDiurnalCycle

For exporting data and results see help_export.

A complete list of REddyProc functions be viewed by clicking on the Index link at the bottom of this help page.

Also have a look at the package vignettes.

Author

Department for Biogeochemical Integration at MPI-BGC, Jena, Germany

References

Reichstein M, Falge E, Baldocchi D et al. (2005) On the separation of net ecosystem exchange into assimilation and ecosystem respiration: review and improved algorithm. Global Change Biology, 11, 1424-1439.