help export

Overview of functions helping with exporting Data and Results

help_export()

Details

Functions helping with exporting data

Export Input data from REddyProc class: sEddyProc_sExportData

Export Computed results from REddyProc class: sEddyProc_sExportResults

Write data frame to ASCII tab-separated text file: fWriteDataframeToFile

Write data frame to ASCII comma-separated text file with units in header row: fWriteFrench23

Writing a file that can be supplied to the REddyProc webservice at MPI-BGC Jena can be done by exporting data from REddyProc class EProc.

df <- EProc$sExportData()
fWriteDataframeToFile(df, "myfilename.txt", isSplitDatetime = TRUE)

For preparing minimal working examples also consider

Omit records before the start of the first full day and the end of the last full day: df <- filter_entire_days(df)

Subset data.frame to one or two years: df <- filter_years_eop(df, c(1998))

There are several functions that import from file of different formats.

Load text file with one header and one unit row into data frame: fLoadTXTIntoDataframe

Reads sequence of annual files in the format of Europe-fluxdata 2016: fLoadEuroFlux16

Read basic variables from Ameriflux standard (as of 2022) files: fLoadAmeriflux22

Read NetCDF files -> moved to separate package REddyProcNCDF (https://github.com/bgctw/REddyProcNCDF)

Back to REddyProc-package.

Author

TW Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav <U+0160>igut <sigut.l@czechglobe.cz> [ctb]