partGLControlLasslopCompatible

Daytime flux partitioning parms compatible with with the pvWave

partGLControlLasslopCompatible(nBootUncertainty = 0L, 
    minNRecInDayWindow = 10L, isAssociateParmsToMeanOfValids = FALSE, 
    isLasslopPriorsApplied = TRUE, isUsingLasslopQualityConstraints = TRUE, 
    isBoundLowerNEEUncertainty = FALSE, fixedTRefAtNightTime = 15, 
    isExtendTRefWindow = FALSE, smoothTempSensEstimateAcrossTime = FALSE, 
    isRefitMissingVPDWithNeglectVPDEffect = FALSE, 
    minPropSaturation = NA, isNeglectVPDEffect = FALSE, 
    replaceMissingSdNEEParms = c(NA, NA), 
    neglectNEEUncertaintyOnMissing = TRUE, 
    ...)

Arguments

nBootUncertainty

0: Derive uncertainty from curvature of a single fit, neglecting the uncertainty of previously estimated temperature sensitivity, E0

minNRecInDayWindow

Minimum number of 10 valid records for regression in a single window

isAssociateParmsToMeanOfValids

associate parameters to the first record of the window for interpolation instead of mean across valid records inside a window

isLasslopPriorsApplied

Apply fixed Lasslop priors in LRC fitting.

isUsingLasslopQualityConstraints

avoid quality constraints additional to the ones in Lasslop 2010

isBoundLowerNEEUncertainty

FALSE: avoid adjustment of very low uncertainties before day-Time fitting that avoids the high leverage those records with unreasonable low uncertainty.

fixedTRefAtNightTime

use fixed (degree Centigrade) temperature sensitivity instead of median data temperature as reference temperature in estimation of temperature sensitivity from night data

isExtendTRefWindow

avoid successively extending the night-time window in order to estimate a temperature sensitivity where previous estimates failed

smoothTempSensEstimateAcrossTime

FALSE: use independent estimates of temperature sensitivity on each windows instead of a vector of E0 that is smoothed over time

isRefitMissingVPDWithNeglectVPDEffect

FALSE: avoid repeating estimation with isNeglectVPDEffect = TRUE

minPropSaturation

NA: avoid quality constraint of sufficient saturation in data This option is overruled, i.e. not considered, if option isUsingLasslopQualityConstraints = TRUE.

isNeglectVPDEffect

FALSE: do not neglect VPD effect

replaceMissingSdNEEParms

do not replace missing NEE, but see option

neglectNEEUncertaintyOnMissing

if there are records with missing uncertainty of NEE inside one window, set all sdNEE to 1. This overrules option replaceMissingSdNEEParms.

...

further arguments to partGLControl

Author

TW Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav <U+0160>igut <sigut.l@czechglobe.cz> [ctb]

See also

Examples

partGLControlLasslopCompatible()
#> $LRCFitConvergenceTolerance
#> [1] 0.001
#> 
#> $nLRCFitConvergenceTolerance
#> [1] 0.001
#> 
#> $nBootUncertainty
#> [1] 0
#> 
#> $minNRecInDayWindow
#> [1] 10
#> 
#> $isAssociateParmsToMeanOfValids
#> [1] FALSE
#> 
#> $isLasslopPriorsApplied
#> [1] TRUE
#> 
#> $isUsingLasslopQualityConstraints
#> [1] TRUE
#> 
#> $isSdPredComputed
#> [1] TRUE
#> 
#> $isFilterMeteoQualityFlag
#> [1] FALSE
#> 
#> $isBoundLowerNEEUncertainty
#> [1] FALSE
#> 
#> $fixedTRefAtNightTime
#> [1] 15
#> 
#> $isExtendTRefWindow
#> [1] FALSE
#> 
#> $smoothTempSensEstimateAcrossTime
#> [1] FALSE
#> 
#> $isNeglectPotRadForNight
#> [1] FALSE
#> 
#> $isNeglectVPDEffect
#> [1] FALSE
#> 
#> $isRefitMissingVPDWithNeglectVPDEffect
#> [1] FALSE
#> 
#> $fixedTempSens
#>   E0 sdE0 RRef
#> 1 NA   NA   NA
#> 
#> $replaceMissingSdNEEParms
#> [1] NA NA
#> 
#> $neglectNEEUncertaintyOnMissing
#> [1] TRUE
#> 
#> $minPropSaturation
#> [1] NA
#> 
#> $useNightimeBasalRespiration
#> [1] FALSE
#>