Estimate bounded temperature sensitivity E0 and RRef of ecosystem respiration

Estimate bounded temperature sensitivity E0 and RRef of ecosystem respiration

partGLEstimateTempSensInBoundsE0Only(
  REco,
  temperatureKelvin,
  prevE0 = NA,
  TRefFit = median(temperatureKelvin, na.rm = TRUE)
)

Arguments

REco

night time NEE

temperatureKelvin

temperature in Kelvin

prevE0

numeric scalar: the previous guess of Temperature Sensitivity

TRefFit

numeric scalar reference temperature for which RRef is estimated. Set it to the center of the data (the default) for best fit (lowest uncertainty)

Value

list with entries

E0 : estimated temperature sensitivty E0 bounded to 50, 400

sdE0 : standard deviation of E0

TRefFit : reference temperature used in the E0 fit

RRefFit : respiration at TRefFit

Details

Basal respiration is reported for temperature of 15 degree Celsius. However during the fit a reference temperature of the median of the dataset is used. This is done to avoid strong correlations between estimated parameters E0 and RRef, that occure if reference temperature is outside the center of the data.

See also