partGLExtractStandardData

Relevant columns from original input with defined names

partGLExtractStandardData(ds, NEEVar = paste0("NEE", 
    suffixDash, "_f"), QFNEEVar = if (!missing(QFNEEVar.s)) QFNEEVar.s else paste0("NEE", 
    suffixDash, "_fqc"), QFNEEValue = if (!missing(QFNEEValue.n)) QFNEEValue.n else 0, 
    NEESdVar = if (!missing(NEESdVar.s)) NEESdVar.s else paste0("NEE", 
        suffixDash, "_fsd"), TempVar = paste0("Tair_f"), 
    QFTempVar = if (!missing(QFTempVar.s)) QFTempVar.s else paste0("Tair_fqc"), 
    QFTempValue = if (!missing(QFTempValue.n)) QFTempValue.n else 0, 
    VPDVar = if (!missing(VPDVar.s)) VPDVar.s else paste0("VPD_f"), 
    QFVPDVar = if (!missing(QFVPDVar.s)) QFVPDVar.s else paste0("VPD_fqc"), 
    QFVPDValue = if (!missing(QFVPDValue.n)) QFVPDValue.n else 0, 
    RadVar = if (!missing(RadVar.s)) RadVar.s else "Rg_f", 
    QFRadVar = if (!missing(QFRadVar.s)) QFRadVar.s else paste0("Rg_fqc"), 
    QFRadValue = if (!missing(QFRadValue.n)) QFRadValue.n else 0, 
    PotRadVar = if (!missing(PotRadVar.s)) PotRadVar.s else "PotRad_NEW", 
    suffix = if (!missing(Suffix.s)) Suffix.s else "", 
    NEEVar.s, QFNEEVar.s, QFNEEValue.n, NEESdVar.s, 
    TempVar.s, QFTempVar.s, QFTempValue.n, 
    VPDVar.s, QFVPDVar.s, QFVPDValue.n, RadVar.s, 
    QFRadVar.s, QFRadValue.n, PotRadVar.s, 
    Suffix.s, controlGLPart = partGLControl())

Arguments

ds

dataset with all the specified input columns and full days in equidistant times

NEEVar

Variable of NEE

QFNEEVar

Quality flag of variable

QFNEEValue

Value of quality flag for _good_ (original) data

NEESdVar

Variable of standard deviation of net ecosystem fluxes

TempVar

Filled air or soil temperature variable (degC)

QFTempVar

Quality flag of filled temperature variable

QFTempValue

Value of temperature quality flag for _good_ (original) data

VPDVar

Filled Vapor Pressure Deficit, VPD (hPa)

QFVPDVar

Quality flag of filled VPD variable

QFVPDValue

Value of VPD quality flag for _good_ (original) data

RadVar

Filled radiation variable

QFRadVar

Quality flag of filled radiation variable

QFRadValue

Value of radiation quality flag for _good_ (original) data

PotRadVar

Variable name of potential rad. (W / m2)

suffix

string inserted into column names before identifier for NEE column defaults (see sEddyProc_sMDSGapFillAfterUstar).

NEEVar.s

deprecated

QFNEEVar.s

deprecated

QFNEEValue.n

deprecated

NEESdVar.s

deprecated

TempVar.s

deprecated

QFTempVar.s

deprecated

QFTempValue.n

deprecated

VPDVar.s

deprecated

QFVPDVar.s

deprecated

QFVPDValue.n

deprecated

RadVar.s

deprecated

QFRadVar.s

deprecated

QFRadValue.n

deprecated

PotRadVar.s

deprecated

Suffix.s

deprecated

controlGLPart

further default parameters, see partGLControl

Details

The LRC fit usually weights NEE records by its uncertainty. In order to also use records with missing NEESdVar, uncertainty of the missing values is by default set to a conservatively high value, parameterized by controlGLPart$replaceMissingSdNEEParms). Controlled by argument replaceMissingSdNEEParms in partGLControl, but overruled by argument neglectNEEUncertaintyOnMissing.

Value

a data.frame with columns

sDateTime

first column of ds, usually the time stamp not used, but usually first column is a DateTime is kept for aiding debug

NEE

NEE filtered for quality flay

sdNEE

standard deviation of NEE with missing values replaced

Temp

Temperature, quality filtered if isTRUE(controlGLPart$isFilterMeteoQualityFlag)

VPD

Water pressure deficit, quality filtered if isTRUE(controlGLPart$isFilterMeteoQualityFlag)

Rg

Incoming radiation

isDay

Flag that is true for daytime records

isNight

Flag that is true for nighttime records

Author

Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav <U+0160>igut <sigut.l@czechglobe.cz> [ctb]