Estimate Rectangular Hyperbolic Light Response Curve parameters

Estimate Rectangular Hyperbolic Light Response Curve parameters

partGLFitLRCOneWindow(
  ds,
  winInfo,
  prevRes,
  E0Win,
  controlGLPart = partGLControl(),
  lrcFitter = RectangularLRCFitter()
)

Arguments

ds

data.frame with numeric columns NEE, sdNEE, Temp (degC), VPD, Rg, and logical columns isNight and isDay

winInfo

one-row data.frame with window information, including iWindow

prevRes

component prevRes from previous result, here with item prevE0

E0Win

data.frame with columns E0, sdE0, RRef from nighttime, one row for each window

controlGLPart

list of further default parameters

lrcFitter

R5 class instance responsible for fitting the light response curve

Value

a list with components resOpt, summary, isValid

Details

Estimation as in Lasslop et al., 2010 for successive periods, i.e. windows. Fitting is done on a subset of the data where isDay, NEE, sdNEE, Temp, Rg, and VPD are all non-NA and isDay is TRUE

See also