sEddyProc sEstimateUstarScenarios

Estimate the distribution of u* threshold by bootstrapping over data

sEddyProc_sEstimateUstarScenarios(ctrlUstarEst = usControlUstarEst(), 
    ctrlUstarSub = usControlUstarSubsetting(), 
    UstarColName = "Ustar", NEEColName = "NEE", 
    TempColName = "Tair", RgColName = "Rg", 
    ..., seasonFactor = usCreateSeasonFactorMonth(sDATA$sDateTime), 
    nSample = 200L, probs = c(0.05, 0.5, 
        0.95), isVerbose = TRUE, suppressWarningsAfterFirst = TRUE)

Arguments

ctrlUstarEst

control parameters for estimating uStar on a single binned series, see usControlUstarEst

ctrlUstarSub

control parameters for subsetting time series (number of temperature and Ustar classes ...), see usControlUstarSubsetting

UstarColName

column name for UStar

NEEColName

column name for NEE

TempColName

column name for air temperature

RgColName

column name for solar radiation for omitting night time data

...

further arguments to sEddyProc_sEstUstarThreshold

seasonFactor

factor of seasons to split (data is resampled only within the seasons)

nSample

the number of repetitions in the bootstrap

probs

the quantiles of the bootstrap sample to return. Default is the 5%, median and 95% of the bootstrap

isVerbose

set to FALSE to omit printing progress

suppressWarningsAfterFirst

set to FALSE to show also warnings for all bootstrap estimates instead of only the first bootstrap sample

Details

The choice of the criterion for sufficiently turbulent conditions (u * > chosen threshold) introduces large uncertainties in calculations based on gap-filled Eddy data. Hence, it is good practice to compare derived quantities based on gap-filled data using a range of u * threshold estimates.

This method explores the probability density of the threshold by repeating its estimation on a bootstrapped sample. By default it returns the 90% confidence interval (argument probs). For larger intervals the sample number need to be increased (argument probs).

Value

updated class. Request results by sEddyProc_sGetEstimatedUstarThresholdDistribution

Author

TW Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav <U+0160>igut <sigut.l@czechglobe.cz> [ctb]

See also