sEddyProc sMDSGapFill

MDS gap filling algorithm adapted after the PV-Wave code and paper by Markus Reichstein.

sEddyProc_sMDSGapFill(Var = Var.s, QFVar = if (!missing(QFVar.s)) QFVar.s else "none", 
    QFValue = if (!missing(QFValue.n)) QFValue.n else NA_real_, 
    V1 = if (!missing(V1.s)) V1.s else "Rg", 
    T1 = if (!missing(T1.n)) T1.n else 50, 
    V2 = if (!missing(V2.s)) V2.s else "VPD", 
    T2 = if (!missing(T2.n)) T2.n else 5, 
    V3 = if (!missing(V3.s)) V3.s else "Tair", 
    T3 = if (!missing(T3.n)) T3.n else 2.5, 
    FillAll = if (!missing(FillAll.b)) FillAll.b else TRUE, 
    isVerbose = if (!missing(Verbose.b)) Verbose.b else TRUE, 
    suffix = if (!missing(Suffix.s)) Suffix.s else "", 
    minNWarnRunLength = if (Var == "NEE") 4 * 
        .self$sINFO$DTS/24 else NA_integer_, 
    Var.s, QFVar.s, QFValue.n, V1.s, T1.n, 
    V2.s, T2.n, V3.s, T3.n, FillAll.b, Verbose.b, 
    Suffix.s, method = "Reichstein05")

Arguments

Var

Variable to be filled

QFVar

Quality flag of variable to be filled

QFValue

Value of quality flag for _good_ (original) data, other data is set to missing

V1

Condition variable 1 (default: Global radiation 'Rg' in W m-2)

T1

Tolerance interval 1 (default: 50 W m-2)

V2

Condition variable 2 (default: Vapour pressure deficit 'VPD' in hPa)

T2

Tolerance interval 2 (default: 5 hPa)

V3

Condition variable 3 (default: Air temperature 'Tair' in degC)

T3

Tolerance interval 3 (default: 2.5 degC)

FillAll

Fill all values to estimate uncertainties

isVerbose

Print status information to screen

suffix

String suffix needed for different processing setups on the same dataset (for explanations see below)

minNWarnRunLength

scalar integer: warn if number of subsequent numerically equal values exceeds this number. Set to Inf or NA for no warnings. defaults for "NEE" to records across 4 hours and no warning for others.

Var.s

deprecated

QFVar.s

deprecated

QFValue.n

deprecated

V1.s

deprecated

T1.n

deprecated

V2.s

deprecated

T2.n

deprecated

V3.s

deprecated

T3.n

deprecated

FillAll.b

deprecated

Verbose.b

deprecated

Suffix.s

deprecated

method

specify "Vekuri23" to use the skewness-bias reducing variant

Details

Value

Gap filling results in sTEMP data frame (with renamed columns).

References

Reichstein, M. et al. (2005) On the separation of net ecosystem exchange into assimilation and ecosystem respiration: review and improved algorithm. Global Change Biology, 11, 1424-1439.

Author

AMM, TW Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav <U+0160>igut <sigut.l@czechglobe.cz> [ctb]