sEddyProc sMDSGapFillUStarScens

gapfilling for several filters of estimated friction velocity Ustar thresholds.

sEddyProc_sMDSGapFillUStarScens(...)

Arguments

...

other arguments to sEddyProc_sMDSGapFillAfterUstar and sEddyProc_sMDSGapFill such as fluxVar

Details

sEddyProc$sMDSGapFillUStarDistr: calling sEddyProc_sMDSGapFillAfterUstar for several filters of friction velocity Ustar.

The scenarios need to be set before by sEddyProc_sSetUstarScenarios or accepting the defaults annual estimates of link{sEddyProc_sEstimateUstarScenarios}.

Then the difference between output columns NEE_U05_f and NEE_U95_f corresponds to the uncertainty introduced by the uncertain estimate of the u* threshold.

Value

Matrix (columns correspond to u* Scenarios) with quality flag from filtering ustar (0 - good data, 1 - filtered data)

Gap filling results in sTEMP data frame (with renamed columns), that can be retrieved by sEddyProc_sExportResults. Each of the outputs is calculated for several u* r-estimates and distinguished by a suffix after the variable. E.g. with an an entry "U05" in uStarSuffixes in sEddyProc_sSetUstarScenarios the corresponding filled NEE can be found in output column "NEE_U05_f".

Author

TW Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav <U+0160>igut <sigut.l@czechglobe.cz> [ctb]

See also