usControlUstarEst

Default list of parameters for determining UStar of a single binned series

usControlUstarEst(ustPlateauFwd = 10, ustPlateauBack = 6, 
    plateauCrit = 0.95, corrCheck = 0.5, 
    firstUStarMeanCheck = 0.2, isOmitNoThresholdBins = TRUE, 
    isUsingCPTSeveralT = FALSE, isUsingCPT = FALSE, 
    minValidUStarTempClassesProp = 0.2, minValidBootProp = 0.4, 
    minNuStarPlateau = 3L)

Arguments

ustPlateauFwd

number of subsequent uStar bin values to compare to in fwd mode

ustPlateauBack

number of subsequent uStar bin values to compare to in back mode

plateauCrit

significant differences between a uStar value and the mean of a "plateau"

corrCheck

threshold value for correlation between Tair and u * data

firstUStarMeanCheck

if first uStar bin average of a class is already larger than this value, the temperature class is skipped.

isOmitNoThresholdBins

if TRUE, bins where no threshold was found are ignored. Set to FALSE to report highest uStar bin for these cases

isUsingCPTSeveralT

set to TRUE to use change point detection without binning uStar but with additionally changed aggregation scheme for several temperature classifications

isUsingCPT

set to TRUE to use change point detection without binning uStar before in usual aggregation method (good for comparing methods, but not recommended, overruled by isUsingCPTSeveralT = TRUE)

minValidUStarTempClassesProp

seasons, in which only less than this proportion of temperature classes a threshold was detected, are excluded from aggregation

minValidBootProp

minimum proportion of bootstrap samples for which a threshold was detected. Below this proportion NA quantiles are reported.

minNuStarPlateau

minimum number of records in plateau, threshold must be larger than mean of this many bins

Author

TW Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav <U+0160>igut <sigut.l@czechglobe.cz> [ctb]

See also

Examples

usControlUstarEst()
#> $ustPlateauFwd
#> [1] 10
#> 
#> $ustPlateauBack
#> [1] 6
#> 
#> $plateauCrit
#> [1] 0.95
#> 
#> $corrCheck
#> [1] 0.5
#> 
#> $firstUStarMeanCheck
#> [1] 0.2
#> 
#> $isOmitNoThresholdBins
#> [1] TRUE
#> 
#> $isUsingCPT
#> [1] FALSE
#> 
#> $isUsingCPTSeveralT
#> [1] FALSE
#> 
#> $minValidUStarTempClassesProp
#> [1] 0.2
#> 
#> $minValidBootProp
#> [1] 0.4
#> 
#> $minNuStarPlateau
#> [1] 3
#>