usControlUstarSubsetting

Default list of parameters for subsetting the data for uStarThreshold estimation

usControlUstarSubsetting(taClasses = 7, UstarClasses = 20, 
    swThr = 10, minRecordsWithinTemp = 100, 
    minRecordsWithinSeason = 160, minRecordsWithinYear = 3000, 
    isUsingOneBigSeasonOnFewRecords = TRUE)

Arguments

taClasses

set number of air temperature classes

UstarClasses

set number of Ustar classes

swThr

nighttime data threshold for solar radiation [Wm-2]

minRecordsWithinTemp

integer scalar: the minimum number of Records within one Temperature-class

minRecordsWithinSeason

integer scalar: the minimum number of Records within one season

minRecordsWithinYear

integer scalar: the minimum number of Records within one year

isUsingOneBigSeasonOnFewRecords

boolean scalar: set to FALSE to avoid aggregating all seasons on too few records

Author

TW Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav <U+0160>igut <sigut.l@czechglobe.cz> [ctb]

See also

Examples

usControlUstarSubsetting()
#> $taClasses
#> [1] 7
#> 
#> $UstarClasses
#> [1] 20
#> 
#> $swThr
#> [1] 10
#> 
#> $minRecordsWithinTemp
#> [1] 100
#> 
#> $minRecordsWithinSeason
#> [1] 160
#> 
#> $minRecordsWithinYear
#> [1] 3000
#> 
#> $isUsingOneBigSeasonOnFewRecords
#> [1] TRUE
#>